// src/app/models/member.model.ts
export class Member {
    id: string;
    uid: string;
    displayName: string;
    email: string;
    emailVerified: boolean;
    photoURL: string;
    username: string;
    avatar: string;
    city: string;
    country: string;
    about: string;
    dob: string;
    skill: string;
    language: string;
    starsign: string;
    gender: string;
    religion: string;
}